+++ test/fixtures/steps.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Steps
+  def upto(n)
+    (1..n).to_a
+  end
+end